==> Scripts/12_DCA_Denoise data for WGCNA.ipynb <==
## Denoising scRNA-seq using DCA 

### Generate AnnData object

import warnings
import numpy as np
import matplotlib.pyplot as plt
import scanpy as sc
import pandas as pd
import anndata
from dca.api import dca
warnings.simplefilter(action="ignore", category = FutureWarning)

# Read in the cleaned data set when genes have been filtered to only keep those with total expression across the board > 1
data = pd.read_csv("output/Competition_forWGCNA.csv", index_col=0)
cellinfo = pd.read_csv("output/DCA_cellInfo.csv", index_col=0)
geneinfo = pd.read_csv("output/DCA_geneInfo.csv", index_col=0)

data_T = data.transpose()

print(data_T.head())
print(data_T.shape)

adata = sc.AnnData(data_T, obs=cellinfo, var = geneinfo)
adata.obs_names = cellinfo.X
adata.var_names = geneinfo.Gene
sc.pp.filter_genes(adata, min_counts = 1)
print(adata)

### Denoise data

%%time
adata_ae=adata.copy()
dca(adata_ae, threads = 1, mode="denoise")

for i in [adata, adata_ae]:
    sc.pp.normalize_per_cell(i)
    sc.pp.log1p(i)
    sc.pp.pca(i)

print(adata)

df_ae = pd.DataFrame(data=adata_ae.X.transpose(), index = adata_ae.var_names, columns=adata_ae.obs_names)

df_ae.head()

df_ae.to_csv("output/Denoised_Competition_matrix.csv")

## END OF FILE
