==> Scripts/13_WGCNA_runTOM.Rmd <==
---
title: 'WGCNA on the GPC subset'
author: "Nguyen Huynh"

date: "12/20/2022"
output: github_document
vignette: >
  %\VignetteEngine{knitr::rmarkdown}
  %\VignetteEncoding[utf8]{inputenc}
abstract: "Required inputs for this script include: "
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, results = TRUE, eval = FALSE)
```

```{r}
library(WGCNA)
options(stringsAsFactors = FALSE)
```

```{r}
dat <- read.csv("output/Denoised_Competition_matrix.csv", row.names = 1)
# Filter for GPC 
meta <- read.csv("output/DCA_cellInfo.csv", header = TRUE, row.names = 1, stringsAsFactors = FALSE)
all.equal(colnames(dat) ,row.names(meta))
dat <- dat[, colnames(dat) %in% row.names(meta)[which(meta$CellType == "GPC")]]
# Use all data
filt <- log2(as.matrix(dat)+1e-6)
filt <- t(filt)
```

```{r}
# Pick soft thresholding power
powers <- c(seq(from=1, to=41, by = 2))
sft <- pickSoftThreshold(filt, dataIsExpr = TRUE, powerVector = powers, corFnc = "bicor", corOptions = list(use="p", maxPOutliers=0.1), networkType = "signed", verbose = TRUE)
```

```{r}
# Plot the results
sizeGrWindow(9, 5)
par(mfrow = c(1,2));
cex1 = 0.9;

# Scale-free topology fit index as a function of the soft-thresholding power
plot(sft$fitIndices[,1], -sign(sft$fitIndices[,3])*sft$fitIndices[,2],xlab="Soft Threshold (power)",ylab="Scale Free Topology Model Fit, signed R^2",type="n", main = paste("Scale independence"));
text(sft$fitIndices[,1], -sign(sft$fitIndices[,3])*sft$fitIndices[,2],labels=powers,cex=cex1,col="red");

# Red line corresponds to using an R^2 cut-off
abline(h=0.9,col="red")

# Mean connectivity as a function of the soft-thresholding power
plot(sft$fitIndices[,1], sft$fitIndices[,5],xlab="Soft Threshold (power)",ylab="Mean Connectivity", type="n",main = paste("Mean connectivity"))
text(sft$fitIndices[,1], sft$fitIndices[,5], labels=powers, cex=cex1,col="red")
```

```{r}
softPower = 19
#turn expression into a topological overlap matrix (TOM) to minimize the effects of noise and spurious associations
TOM=TOMsimilarityFromExpr(filt,networkType = "signed", TOMType = "signed", power = softPower, corType="bicor", maxPOutliers = 0.1)

# save.image("output/WGCNA.RData")
colnames(TOM) <- colnames(filt)
row.names(TOM) <- colnames(filt)
dissTOM <- 1- TOM
geneTree <- flashClust::flashClust(as.dist(dissTOM), method = "average")
```

```{r}
minModuleSize <- 20
dynamicMod <- cutreeDynamic(dendro = geneTree, distM = dissTOM, method = "hybrid", minClusterSize = minModuleSize, deepSplit = 4, pamRespectsDendro = FALSE, cutHeight = 0.99, respectSmallClusters = TRUE)
table(dynamicMod)
dynamicColor <- labels2colors(dynamicMod)

MEs <- moduleEigengenes(filt, dynamicColor)
MEG <- MEs$eigengenes

names(dynamicColor) <- colnames(filt)
# save(filt, dynamicColor, MEs, geneTree, file="output/WGCNA_output.RData")
# saveRDS(MEG, "output/MEG.RDS")
```

```{r eval = TRUE}
sessionInfo()
```
