==> Scripts/15_HD_IPA_Network_Construction.Rmd <==
---
title: "Network Construction from SCENIC, WGCNA, and IPA"
author: "John Mariani"
date: "12/6/2022"
output: github_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = "..")
```


```{r, echo = TRUE, message=FALSE, warning=FALSE}

library(data.table)
library(dplyr)
library(tidyr)

```

# Load Network edges from SCENIC and WGCNA analysis

```{r}

networkEdges <- read.delim("output/regulons_filt_module.txt")
networkNodes <- data.frame(Node = unique(c(networkEdges$TF, networkEdges$Target)))
networkNodes$Node <- as.character(networkNodes$Node)


```

# Filter for genes DE in any comparison in either model and export for IPA

```{r}

allDE <- read.delim("DE/allDE.txt")

networkEdges_de <- networkEdges[networkEdges$Target %in% allDE$Gene,]
networkNodes_de <- data.frame(Node = unique(c(networkEdges_de$TF, networkEdges_de$Target)))

write.table(networkNodes_de, "data_for_import/network_de_for_ipa.txt", sep = "\t", quote = F, row.names = F)


```

# Import IPA information and export network edges

```{r}

files <- "data_for_import/ipa_blue_regulon_de.txt"
compNames <- c("TFregulons")
upstreamIPA <- fread(files, skip = "Upstream Regulators")
upstreamIPA <- upstreamIPA[,c(3,10,11)]
names(upstreamIPA) <- c("Pathway", "pVal", "Genes")
upstreamIPA$Pathway <- paste0(upstreamIPA$Pathway, " Signaling")
canonicalIPA <- fread(files, skip = "Canonical Pathways")
canonicalIPA <- canonicalIPA[,c(1,2,5)]
names(canonicalIPA)<- c("Pathway", "pVal", "Genes")
functionalIPA <- fread(files, skip = "Diseases and Bio Functions")
functionalIPA <- functionalIPA[,c(3,4,9)]
names(functionalIPA) <- c("Pathway", "pVal", "Genes")
IPA <- rbind(upstreamIPA, canonicalIPA)
IPA <- rbind(IPA, functionalIPA)

edges <- IPA %>% 
  mutate(Gene = strsplit(as.character(Genes), ",")) %>% 
  unnest(Gene) %>% .[,-3]

edges <- edges[edges$Gene %in% networkNodes$Node,]

goCats <- c("EIF2 Signaling", "LARP1 Signaling", "MLXIPL Signaling", "mTOR Signaling", "MYC Signaling", "MYCN Signaling", "Oxidative Phosphorylation", "RICTOR Signaling", "Synthesis of protein", "TP53 Signaling", "YAP1 Signaling", "CEBPZ Signaling",
            "CTCF Signaling", "E2F1 Signaling", "ETV4 Signaling", "NFYB Signaling")

edges <- edges[edges$Pathway %in% goCats,]

names(edges)

networkEdges_de$Type <- "SCENIC"
names(networkEdges_de)[1] <- "Source"

edges$Pathway <- gsub(x = edges$Pathway, pattern = " Signaling", replacement = "")
edges$pVal <- NULL
names(edges) <- c("Source", "Target")
edges$Type <- "IPA"

scenicEdges <- networkEdges_de
scenicEdges$Weight <- NULL


networkEdges_Final <- rbind(scenicEdges, edges)


write.table(networkEdges_Final, "data_for_import/networkEdges_de.txt", quote = F, row.names = F, sep = "\t")



```


```{r}

sessionInfo()

```
